requirements.txt
# Hub CLI Dependencies
PyYAML>=6.0.1

# Development and Quality
ruff>=0.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
